Add files via upload
Who This Book is For
This book contains only those parts of the vast body of material on machine learning developed
since the 1960s that have proven to have a significant practical value. A beginner in machine
learning will find in this book just enough details to get a comfortable level of understanding
of the field and start asking the right questions.
Practitioners with experience can use this book as a collection of directions for further
self-improvement. The book also comes in handy when brainstorming at the beginning of a
project, when you try to answer the question of whether a given technical or business problem
is “machine-learnable” and, if yes, which techniques you should try to solve it.